backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
python-dotenv
scikit-learn
numpy
scipy
python-levenshtein
